+++ rambo/cli.py
@@ -37,12 +37,13 @@ class ConfigSectionSchema(object):
         ram_size           = Param(type=int)
         drive_size         = Param(type=int)
         machine_type       = Param(type=str)
-        sync_dir           = Param(type=click.Path())
+        sync_dirs           = Param(type=str)
         sync_type           = Param(type=str)
         provision          = Param(type=bool)
         destroy_on_error   = Param(type=bool)
         hostname = Param(type=str)
-        provision_cmd = Param(type=click.Path())
+        project_dir = Param(type=click.Path())
+        provision_cmd = Param(type=str)
         provision_script = Param(type=click.Path())
         provision_with_salt = Param(type=bool)
         salt_bootstrap_args = Param(type=str)
@@ -242,10 +243,12 @@ def ssh_cmd(ctx, command):
               'E.g. m5.medium for ec2, or s-8vcpu-32gb for digitalocean.\n')
 @click.option('--salt-bootstrap-args', type=str,
               help='Args to pass to salt-bootstrap when provisioning with Salt.')
-@click.option('--sync-dir', type=click.Path(resolve_path=True),
-              help='Path to sync into VM')
+@click.option('--sync-dirs', type=str,
+              help='Paths to sync into VM, passed as a Python list of lists of the form "[['source', 'target'], ['source2', 'target2']]".')
 @click.option('--sync-type', type=str,
               help='Sync type')
+@click.option('--project-dir', type=click.Path(resolve_path=True),
+              help='List of path to sync into VM')
 @click.option('--provision/--no-provision', default=None,
               help='Enable or disable provisioning')
 @click.option('-c', '--provision-cmd', type=str,
@@ -260,7 +263,7 @@ def ssh_cmd(ctx, command):
               help='The name of the VirtualMachine / Container.')
 @click.pass_context
 def up_cmd(ctx, provider, box, hostname, guest_os, ram_size, cpus, drive_size, machine_type,
-           salt_bootstrap_args, sync_dir, sync_type, provision, provision_cmd, provision_script, provision_with_salt, destroy_on_error, vm_name):
+           salt_bootstrap_args, sync_dirs, sync_type, project_dir, provision, provision_cmd, provision_script, provision_with_salt, destroy_on_error, vm_name):
     '''Start a VM or container. Will create one and begin provisioning it if
     it did not already exist. Accepts many options to set aspects of your VM.
     Precedence is CLI > Config > Env Var > defaults.